Revert "At least to the point where it works."
This reverts commit 838ed0aada6840693f8c9d9b5a639d3aeb3e6bb0.
